Create Authenticator abstraction and implementations
